switched from int to float32

+++ calc/calc_add.go
@@ -1,6 +1,6 @@
 package calc
 
 // Add takes two numbers and adds them together
-func Add(a int, b int) int {
+func Add(a float32, b float32) float32 {
     return a + b;
 }
