+++ calc/calc_divide.go
@@ -1,6 +1,6 @@
 package calc
 
 // Divide takes two numbers and adds them together
-func Divide(a int, b int) int {
+func Divide(a float32, b float32) float32 {
     return a/b;
 }
